[UPD] Cleaned repo and updated refs
Cleaned repo from bloat folders used for distribution and packaging.
Also added a sanitizer for HTML in the main code, to prevent an error when trying to send messages which cause conflict with the json format.
Finally, updated the server url references in both main code and readme, because previous one was not working anymore; as well as removing the html reference for Telegram in the Readme, since the sanitizer was added.